% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellboro-package.r
\docType{package}
\name{cellboro-package}
\alias{cellboro}
\alias{cellboro-package}
\title{cellboro: For a point, finds its Delaunay triangulation-based neighbours 
and the Voronoi cell area}
\description{
The package simplifies to leverage Delaunay triangulation and the dual
Voronoi diagrams on the cell spatial data. Delaunay triangulation shown 
the adjacent points (cells) for each cell; 
Voronoi domain area shows the area that a cell controls.
After careful selection, we use the `interp` R package for both tasks, 
it is the best for marginal points. If one needs more sophisticated analysis
of the diagram, the `interp` can be used directly, and the results will be
in accordance with the `cellboro` calls. 
We added a function that returns a list of pairwise neignbours,
the original function provided by `interp` was too slow.
}
\author{
\strong{Maintainer}: Alexander Favorov \email{favorov@sensi.org} (\href{https://orcid.org/0000-0003-2595-1091}{ORCID})

Authors:
\itemize{
  \item Ludmila Danilova \email{ludmila.danilova@gmail.com} [contributor]
}

}
\keyword{Delaunay}
\keyword{Voronoi}
\keyword{spatial}
